menambahkan fitur hanya admin dengan rw yang sama yang dapat memperbarui status laporan user

+++ app/Http/Controllers/Admin/ReportStatusController.php
@@ -20,57 +20,49 @@ class ReportStatusController extends Controller
 
     public function __construct(
         ReportRepositoryInterface $reportRepository,
-        ReportStatusRepositoryInterface $reportStatusRepository,
-        )
-    {
+        ReportStatusRepositoryInterface $reportStatusRepository
+    ) {
         $this->reportRepository = $reportRepository;
         $this->reportStatusRepository = $reportStatusRepository;
-    }
-
-    public function index()
-    {
-        
     }
 
     public function create($reportId)
     {
         $report = $this->reportRepository->getReportById($reportId);
+        $this->authorize('manageStatus', $report);
         return view('pages.admin.report-status.create', compact('report'));
     }
 
     public function store(StoreReportStatusRequest $request)
     {
-        $data = $request->validated();
+        $report = $this->reportRepository->getReportById($request->report_id);
+        $this->authorize('manageStatus', $report);
 
+        $data = $request->validated();
         if ($path = $this->handleFileUpload($request, 'image', 'assets/report-status/image')) {
             $data['image'] = $path;
         }
-
         $data['created_by_role'] = 'admin';
 
         $this->reportStatusRepository->createReportStatus($data);
 
         Swal::success('Success', 'Data Progress laporan berhasil ditambahkan!')->timerProgressBar();
-
         return redirect()->route('admin.report.show', $request->report_id);
-    }
-
-    public function show(string $id)
-    {
-        
     }
 
     public function edit(string $id)
     {
         $status = $this->reportStatusRepository->getReportStatusById($id);
-
+        $this->authorize('manageStatus', $status->report);
         return view('pages.admin.report-status.edit', compact('status'));
     }
 
     public function update(UpdateReportStatusRequest $request, string $id)
     {
-        $data = $request->validated();
+        $status = $this->reportStatusRepository->getReportStatusById($id);
+        $this->authorize('manageStatus', $status->report);
 
+        $data = $request->validated();
         if ($path = $this->handleFileUpload($request, 'image', 'assets/report-status/image')) {
             $data['image'] = $path;
         }
@@ -78,18 +70,17 @@ class ReportStatusController extends Controller
         $this->reportStatusRepository->updateReportStatus($data, $id);
 
         Swal::success('Success', 'Data progress laporan berhasil diubah!')->timerProgressBar();
-
         return redirect()->route('admin.report.show', $request->report_id);
     }
 
     public function destroy(string $id)
     {
         $status = $this->reportStatusRepository->getReportStatusById($id);
+        $this->authorize('manageStatus', $status->report);
 
         $this->reportStatusRepository->deleteReportStatus($id);
 
         Swal::success('Success', 'Data progress laporan berhasil dihapus!')->timerProgressBar();
-
         return redirect()->route('admin.report.show', $status->report_id);
     }
 }
